requirements.txt
plottable==0.1.5
python-dotenv>=1.1.0
osu.py>=4.0.2
discord.py>=2.5.2
aiosqlite>=0.21.0
flask>=3.1.0
flask_sqlalchemy>=3.1.1
itsdangerous>=2.2.0
supabase>=2.15.3
pandas>=2.2.3
pytz>=25.1
aiofiles>=24.1.0
python-dateutil>=2.9.0.post0
matplotlib>=3.10.3
quart>=0.20.0
